requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
#Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
openai==0.28.1
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
tiktoken==0.5.1
Flask-Markdown==0.3
markdown2==2.4.10
pypdf2==3.0.1
typing_extensions==4.8.0
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.0
blinker==1.6.3
cachelib==0.10.2
click==8.1.7
flask==3.0.0
flask-session==0.5.0
itsdangerous==2.1.2
redis==5.0.1